requirements.txt
adminlte-base==0.2.0
apiflask==0.11.0; python_version >= "3.7"
apispec==5.1.1; python_version >= "3.7"
arrow==1.2.2; python_version >= "3.6"
asttokens==2.0.5
bcrypt==3.2.0; python_version >= "3.6"
birdseye==0.9.3
bootstrap-flask==2.0.1; python_version >= "3.7"
brotli==1.0.9; python_version >= "3.7"
cached-property==1.5.2
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0"
cachetoolsutils==1.1.0
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
cffi==1.15.0; python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3"
cheap-repr==0.5.1
click==8.0.3; python_version >= "3.7"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0"
djask==0.3.2; python_version >= "3.7"
dnspython==2.2.0; python_version >= "3.6" and python_version < "4.0"
docopt==0.6.2
email-validator==1.1.3; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.5.0"
flask-admin==1.6.0; python_version >= "3.6"
flask-adminlte-full==0.2.0
flask-basicauth==0.2.0
flask-bcrypt==0.7.1
flask-compress==1.10.1; python_version >= "3.7"
flask-gravatar==0.5.0
flask-httpauth==4.5.0; python_version >= "3.7"
flask-humanize==0.3.0
flask-login==0.5.0
flask-marshmallow==0.14.0; python_version >= "3.7"
flask-mongoengine==1.0.0
flask-sqlalchemy==2.5.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
flask-wtf==0.15.1; python_version >= "3.7"
flask==2.0.2; python_version >= "3.6"
flasksimpleauth==5.2.0
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") and python_full_version >= "3.5.0"
humanize==4.0.0; python_version >= "3.7"
idna==3.3; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.6.0"
itsdangerous==2.0.1; python_version >= "3.7"
jinja2==3.0.3; python_version >= "3.7"
libsass==0.21.0
littleutils==0.2.2
markupsafe==2.0.1; python_version >= "3.7"
marshmallow-sqlalchemy==0.27.0; python_version >= "3.7"
marshmallow==3.14.1; python_version >= "3.7"
mongoengine==0.23.1; python_version >= "3.6"
npx==0.0.25; python_version >= "3.7"
numpy==1.22.1; python_version >= "3.8"
outdated==0.2.1
packaging==21.3; python_version >= "3.7"
pipreqs==0.4.11
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pymongo==3.9.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pyparsing==3.0.7; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python-dotenv==0.19.2; python_version >= "3.5"
requests==2.27.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
sqlalchemy==1.4.31; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
traceback-with-variables==2.0.4; python_version >= "3.5"
urllib3==1.26.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
webargs==8.1.0; python_version >= "3.7"
werkzeug==2.0.2; python_version >= "3.7"
wtforms-sqlalchemy==0.3; python_version >= "3.7"
wtforms==3.0.1; python_version >= "3.7"
yarg==0.1.9
